[polygamma] Fix signed/unsigned comparison warning

+++ include/boost/math/special_functions/detail/polygamma.hpp
@@ -396,7 +396,7 @@
      // However, if the final evaluation would have too many
      // terms just bail out right away:
      //
-     if(n / 2 > policies::get_max_series_iterations<Policy>())
+     if((unsigned)n / 2u > policies::get_max_series_iterations<Policy>())
         return policies::raise_evaluation_error<T>(function, "The value of n is so large that we're unable to compute the result in reasonable time, best guess is %1%", 0, pol);
 #ifdef BOOST_HAS_THREADS
      static boost::detail::lightweight_mutex m;
